% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStreamsCpu.R
\name{createStreamsCpu}
\alias{createStreamsCpu}
\title{createStreamsCpu}
\usage{
createStreamsCpu(n, initial = sample.int(2147483647, 6, replace = TRUE))
}
\arguments{
\item{n}{Number of streams to create}

\item{initial}{Initial state of the first stream, length 6, recycled if shorter}
}
\value{
A stream object on CPU
}
\description{
Create streams stored on the CPU
}
\author{
Ruoyong Xu
}
