% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStreamsGpu.R
\name{createStreamsGpu}
\alias{createStreamsGpu}
\title{Create streams on GPU}
\usage{
createStreamsGpu(n, initial)
}
\arguments{
\item{n}{number of streams to create.}

\item{seed}{the initial seed of streams.}
}
\value{
A stream object on GPU.
}
\description{
Create streams in R.
}
