% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemInfo.R
\name{get_system_info}
\alias{get_system_info}
\title{get_system_info}
\usage{
get_system_info()
}
\value{
List of system information.
}
\description{
This function prints operating system, CPU, GPU, and OpenCL version information.
}
