% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{month}
\alias{month}
\title{Sample dataset: month.RData}
\format{
A data frame with 12 rows and 12 columns:
\describe{
    \item{row}{Month of the year}
    \item{column}{Congenital anomalies category of the newborn}
  }
}
\source{
https://www.cdc.gov/nchs/data_access/VitalStatsOnline.htm
}
\usage{
data(month)
}
\description{
Counts of birth anomalies by birth month and congenital 
anomalies categories in 2018 within the United States.
}
