% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemInfo.R
\name{print_system_info}
\alias{print_system_info}
\title{print_system_info}
\usage{
print_system_info()
}
\value{
A character vector containing system information.
}
\description{
This function prints operating system, CPU, GPU, and OpenCL version information.
}
