% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp.R
\name{rexp}
\alias{rexp}
\title{rexp}
\usage{
rexp(n, rate = 1, streams, Nglobal, type = c("double", "float"))
}
\arguments{
\item{n}{a number or a vector specifying the size of output matrix}

\item{rate}{mean equals to 1/rate}

\item{streams}{streams}

\item{Nglobal}{vector of length 2, number of work items}

\item{type}{random number type, one of "uniform" or "normal"}
}
\value{
a vclVector or vclMatrix of exponential random numbers
}
\description{
rexp
}
