% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexpGpu.R
\name{rexpGpu}
\alias{rexpGpu}
\title{rexpGpu}
\usage{
rexpGpu(n, rate = 1, streams, Nglobal = NULL, type = NULL, verbose = NULL)
}
\arguments{
\item{n}{A number or a vector specifying the size of the output vector or matrix}

\item{rate}{Distribution parameter, mean equals to 1/rate}

\item{streams}{a vclMatrix of streams.}

\item{Nglobal}{NDRange of work items for use}

\item{type}{"double" or "float" of generated random numbers}

\item{verbose}{if TRUE, print extra information}
}
\value{
a 'vclVector' or 'vclMatrix' of exponential random numbers
}
\description{
Generate exponential random numbers parallely on a GPU
}
\details{
\code{type} specifies the precision type of random numbers. If GPU supports "double", default is "double", otherwise, default is "single"
}
\examples{
library('clrng')
setContext(grep("gpu", listContexts()$device_type)[1])
streams <- createStreamsGpu(8)
as.vector(rexpGpu(7, streams=streams))

change to produce float precision random numbers
options(type='float')
as.matrix(rexpGpu(c(2,3), rate=0.5, streams))

}
