% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm.R
\name{rnorm}
\alias{rnorm}
\title{rnorm}
\usage{
rnorm(
  n,
  streams,
  Nglobal,
  type = c("float", "double")[1 + gpuInfo()$double_support],
  verbose = FALSE
)
}
\arguments{
\item{n}{A number or a vector specifying the size of output vector or matrix}

\item{streams}{Streams object}

\item{Nglobal}{NDRange of work items for use}

\item{type}{Precision type of random numbers, "double" or "float"}

\item{verbose}{if TRUE, print extra information}
}
\value{
A 'vclVector' or 'vclMatrix' of standard normal random numbers
}
\description{
Generate standard normal random numbers on a GPU
}
