% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnumber.R
\name{rnumberGpu}
\alias{rnumberGpu}
\title{runif gpu}
\usage{
rnumberGpu(x, streams, workgroupSize, localSize = c(2, 2),
  random_type = c("uniform", "normal"))
}
\arguments{
\item{x}{vector}

\item{streams}{vector of integers, 18x workgroupSize}

\item{workgroupSize}{workgroup size}

\item{localSize}{local size}
}
\value{
altered in-place
}
\description{
Random uniforms
}
