% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif.R
\name{runif}
\alias{runif}
\title{runif}
\usage{
runif(
  n,
  streams,
  Nglobal,
  type = c("float", "double", "integer")[1 + gpuInfo()$double_support],
  verbose = FALSE
)
}
\arguments{
\item{n}{A number or a vector specifying the size of output vector or matrix}

\item{streams}{Streams object}

\item{Nglobal}{NDRange of work items for use}

\item{type}{Precision type of random numbers, "double" or "float" or "integer", default is double}

\item{verbose}{if TRUE, print extra information}
}
\value{
A vclVector or vclMatrix of uniform random numbers
}
\description{
Generate uniform random numbers on a GPU
}
\examples{
 
library(clrng)
as.vector(runif(5, Nglobal=c(4,2)))
as.matrix(runif(c(2,2), Nglobal=c(2,2), type="float"))
}
