% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runif.R
\name{runifGpu}
\alias{runifGpu}
\title{runif gpu}
\usage{
runifGpu(x, workgroupSize = 1, streams, localSize = 1)
}
\arguments{
\item{x}{vector}

\item{workgroupSize}{workgroup size}

\item{streams}{vector of integers, 18x workgroupSize}

\item{localSize}{local size}
}
\value{
altered in-place
}
\description{
Random uniforms
}
