% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStreamsGpu.R
\name{setBaseCreator}
\alias{setBaseCreator}
\title{setBaseCreator}
\usage{
setBaseCreator(initial = rep(12345, 6))
}
\arguments{
\item{initial}{a vector of six 31-bit integers specifying the initial state of the first stream. Default is c(12345,12345,12345,12345,12345,12345).}
}
\value{
a stream object of class 'vclMatrix'.
}
\description{
Set the initial seed of the first stream.
}
\details{
\code{initial} is of length 6, recycled if shorter.
}
\examples{
library('clrng')
setBaseCreator(c(111,222,333,444,555,666))
}
