% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{week}
\alias{week}
\title{Sample dataset: week.RData}
\format{
A data frame with 7 rows and 12 columns:
\describe{
    \item{row}{Day of the week}
    \item{column}{Congenital anomalies category of the newborn}
  }
}
\source{
https://www.cdc.gov/nchs/data_access/VitalStatsOnline.htm
}
\usage{
data(week)
}
\description{
Frequencies for birth anomalies by birth day of week and 
congenital anomalies categories in 2018 within the United States.
}
