% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStreamsCpu.R
\name{createStreamsCpu}
\alias{createStreamsCpu}
\title{createStreamsCpu}
\usage{
createStreamsCpu(n, initial = NULL)
}
\arguments{
\item{n}{Number of streams to create}

\item{initial}{Initial state of the first stream, length 6, recycled if shorter}
}
\value{
A stream object on CPU
}
\description{
Create streams stored on the CPU
}
\examples{
library(clrng)
library(gpuR)
t(createStreamsCpu(n=4, initial=12345))
###GPU streams
myStreamsGpu1 = vclMatrix(createStreamsCpu(n=4, initial=12345))

}
\author{
Ruoyong Xu
}
