% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStreamsCpu.R
\name{createStreamsCpu}
\alias{createStreamsCpu}
\title{createStreamsCpu}
\usage{
createStreamsCpu(n, initial)
}
\arguments{
\item{n}{number of streams to create}

\item{initial}{initial state of first stream, length 6, recycled if shorter}
}
\value{
A stream object on CPU.
}
\description{
create streams stored on the CPU.
}
