% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createStreamsCpu.R
\name{createStreamsCpu}
\alias{createStreamsCpu}
\title{createStreamsCpu}
\usage{
createStreamsCpu(n)
}
\arguments{
\item{n}{Number of streams to create}
}
\value{
A stream object on CPU
}
\description{
Create streams stored on the CPU
}
\examples{
library(clrng)
library(gpuR)
t(createStreamsCpu(n=5))
###GPU streams
myStreamsGpu1 = vclMatrix(createStreamsCpu(n=4))

}
