% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStreamsGpu.R
\name{createStreamsGpu}
\alias{createStreamsGpu}
\title{createStreamsGpu}
\usage{
createStreamsGpu(n, initial = sample.int(2147483647, 6, replace = TRUE))
}
\arguments{
\item{n}{number of streams to create}

\item{initial}{Initial state of the first stream, length 6, recycled if shorter}
}
\value{
A stream object of class 'vclMatrix' on GPU
}
\description{
Generate streams on GPU device
}
