% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStreamsGpu.R
\name{createStreamsGpu}
\alias{createStreamsGpu}
\title{createStreamsGpu}
\usage{
createStreamsGpu(n = prod(getOption("clrng.Nglobal")))
}
\arguments{
\item{n}{number of streams to create, default is 1024}
}
\value{
A stream object of class 'vclMatrix' on GPU
}
\description{
Generate streams on GPU device
}
\examples{
setBaseCreator(rep(12345,6))
myStreamsGpu = createStreamsGpu(n=4)
t(as.matrix(myStreamsGpu))
myStreamsGpu2 = createStreamsGpu(n=4)
t(as.matrix(myStreamsGpu2))
}
