% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_sim.R
\name{fisher.sim}
\alias{fisher.sim}
\title{fisher.sim}
\usage{
fisher.sim(
  x,
  N,
  streams,
  type = c("float", "double")[1 + gpuR::gpuInfo()$double_support],
  returnStatistics = FALSE,
  Nglobal,
  verbose = FALSE
)
}
\arguments{
\item{x}{A contingency table, a vclMatrix}

\item{N}{Requested a number of replicates}

\item{streams}{Streams objects}

\item{type}{"double" or "float" of returned test statistics}

\item{returnStatistics}{logical, if TRUE, return test statistics}

\item{Nglobal}{Size of the index space for use}

\item{verbose}{if TRUE, print extra information}
}
\value{
A list of results, including p-value, actual number of replicates, test statistics and so on
}
\description{
Monte carlo's simulation for Fisher's exact test on GPU
}
