% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fisher_sim.R
\name{fisher.sim}
\alias{fisher.sim}
\title{fisher.sim}
\usage{
fisher.sim(
  x,
  N,
  streams,
  type = c("float", "double")[1 + gpuR::gpuInfo()$double_support],
  returnStatistics = FALSE,
  Nglobal,
  verbose = FALSE
)
}
\arguments{
\item{x}{A contingency table, a vclMatrix}

\item{N}{Requested a number of replicates}

\item{streams}{Streams objects. Default using GPU streams with package default initial seeds}

\item{type}{"double" or "float" of returned test statistics}

\item{returnStatistics}{logical, if TRUE, return test statistics}

\item{Nglobal}{Size of the index space for use}

\item{verbose}{if TRUE, print extra information}
}
\value{
A list of results, including p-value, actual number of replicates, test statistics and so on
}
\description{
Monte carlo's simulation for Fisher's exact test on GPU
}
\examples{
Job <- matrix(c(1,2,1,0, 3,3,6,1, 10,10,14,9, 6,7,12,11), 4, 4)
Job <- gpuR::vclMatrix(Job, type="integer")
streams <- createStreamsGpu(n=64*16, initial=c(1,2,306,6,9,5))
result <- fisher.sim(Job, 1e5, returnStatistics=FALSE, type="double", streams=streams, Nglobal = c(64,16))
result$p.value
result$simNum

}
