% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfactSum.R
\name{logfactSum}
\alias{logfactSum}
\title{logfactsumBackend on GPU}
\usage{
logfactSum(x, Nglobal)
}
\arguments{
\item{x}{an R matrix}

\item{Nglobal}{the size of the index space for use}
}
\value{
the sum of log-factorials of the matrix
}
\description{
logfactsumBackend on GPU
}
