% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logfactSum.R
\name{logfactSum}
\alias{logfactSum}
\title{logfactSum}
\usage{
logfactSum(x, Nglobal)
}
\arguments{
\item{x}{A matrix}

\item{Nglobal}{Size of the index space for use}
}
\value{
Sum of log-factorials of elements of the matrix
}
\description{
Computes the log-factorial of a table
}
