% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{month}
\alias{month}
\title{A dataset stored in x.RData}
\format{
A data frame with columns: Month, Days
}
\source{
Some source or description of where the data came from.
}
\usage{
data(month)
}
\description{
Description of your dataset stored in x.RData.
You can provide more details here.
}
