% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqnorm.R
\name{qqnorm}
\alias{qqnorm}
\title{qqnorm}
\usage{
qqnorm(
  y,
  ylim,
  mu = 0,
  sigma = 1,
  lowertail = 1,
  main = "Normal Q-Q Plot",
  xlab = "Theoretical Quantiles",
  ylab = "Sample Quantiles",
  Nglobal,
  Nlocal = c(2, 2),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{y}{Data sample}

\item{ylim}{Limits on the plot region}

\item{mu}{Mean of Normal distribution, default is 0}

\item{sigma}{Variance of Normal distribution, default is 1}

\item{lowertail}{Logical, whether use lower tail probability, default is TRUE}

\item{main}{plot label}

\item{xlab}{plot label}

\item{ylab}{plot label}

\item{Nglobal}{Size of the index space for use, default is 64 by 4}

\item{Nlocal}{Work group size of the index space}

\item{verbose}{if TRUE, print extra information}

\item{...}{other graphical parameters}
}
\value{
the Normal Q-Q plot
}
\description{
Calculate quantiles on GPU and produce a Q-Q plot in R
}
\examples{
library(clrng)
y <- stats::rt(256, df = 5)
clrng::qqnorm(y, ylim=c(-2,5), mu=0, sigma=1, Nglobal=c(16,4))
x<-stats::rnorm(256)
clrng::qqnorm(x, mu=0, sigma=1, Nglobal=c(8,6))

}
