% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp.R
\name{rexp}
\alias{rexp}
\title{rexp}
\usage{
rexp(
  n,
  rate = 1,
  streams,
  Nglobal,
  type = c("double", "float"),
  verbose = FALSE
)
}
\arguments{
\item{n}{A number or a vector specifying the size of the output vector or matrix}

\item{rate}{Distribution parameter, mean equals to 1/rate}

\item{streams}{Streams object}

\item{Nglobal}{NDRange of work items for use}

\item{type}{"double" or "float" of generated random numbers}
}
\value{
a 'vclVector' or 'vclMatrix' of exponential random numbers
}
\description{
Generate exponential random numbers on a GPU
}
