% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexp.R
\name{rexp}
\alias{rexp}
\title{rexp}
\usage{
rexp(
  n,
  rate = 1,
  streams,
  Nglobal,
  type = c("float", "double")[1 + gpuInfo()$double_support],
  verbose = FALSE
)
}
\arguments{
\item{n}{A number or a vector specifying the size of the output vector or matrix}

\item{rate}{Distribution parameter, mean equals to 1/rate}

\item{streams}{Streams object}

\item{Nglobal}{NDRange of work items for use}

\item{type}{"double" or "float" of generated random numbers}

\item{verbose}{if TRUE, print extra information}
}
\value{
a 'vclVector' or 'vclMatrix' of exponential random numbers
}
\description{
Generate exponential random numbers on a GPU
}
\examples{
library(clrng)
as.vector(rexp(7, Nglobal=c(4,2)))
as.matrix(rexp(c(2,3), rate=0.5, Nglobal=c(2,2), type="float"))
streams <- createStreamsGpu(6, initial = rep(5,6))
as.vector(rnorm(3, streams=streams, Nglobal=c(3,2)))

}
