% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rexpGpu.R
\name{rexpGpu}
\alias{rexpGpu}
\title{rexpGpu}
\usage{
rexpGpu(
  n,
  rate = 1,
  streams,
  Nglobal,
  type = c("float", "double")[1 + gpuInfo()$double_support],
  verbose = FALSE
)
}
\arguments{
\item{n}{A number or a vector specifying the size of the output vector or matrix}

\item{rate}{Distribution parameter, mean equals to 1/rate}

\item{streams}{Streams object. Default using GPU streams with package default initial seeds}

\item{Nglobal}{NDRange of work items for use}

\item{type}{"double" or "float" of generated random numbers}

\item{verbose}{if TRUE, print extra information}
}
\value{
a 'vclVector' or 'vclMatrix' of exponential random numbers
}
\description{
Generate exponential random numbers on a GPU
}
\examples{
library('clrng')
library('gpuR')
streams <- createStreamsGpu(8)
as.vector(rexpGpu(7, streams=streams, Nglobal=c(4,2)))
as.matrix(rexpGpu(c(2,3), rate=0.5, streams, Nglobal=c(4,2), type="float"))

}
