% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnorm.R
\name{rnorm}
\alias{rnorm}
\title{rnorm}
\usage{
rnorm(n, streams, Nglobal, type = c("double", "float"), verbose = FALSE)
}
\arguments{
\item{n}{A number or a vector specifying the size of output vector or matrix}

\item{streams}{Streams object}

\item{Nglobal}{NDRange of work items for use}

\item{type}{Precision type of random numbers, "double" or "float"}

\item{verbose}{print extra information}
}
\value{
A 'vclVector' or 'vclMatrix' of standard normal random numbers
}
\description{
Generate standard normal random numbers on a GPU
}
