% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnumber.R
\name{rnumberGpu}
\alias{rnumberGpu}
\title{runif gpu}
\usage{
rnumberGpu(x, workgroupSize = 1, streams, localSize = 1,
  random_type = c("Uniform", "Normal"))
}
\arguments{
\item{x}{vector}

\item{workgroupSize}{workgroup size}

\item{streams}{vector of integers, 18x workgroupSize}

\item{localSize}{local size}
}
\value{
altered in-place
}
\description{
Random uniforms
}
