% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runifGpu.R
\name{runifGpu}
\alias{runifGpu}
\title{runifGpu}
\usage{
runifGpu(n, streams, Nglobal = NULL, type = NULL, verbose = NULL)
}
\arguments{
\item{n}{a number or numeric vector specifying the size of output vector or matrix}

\item{streams}{a vclMatrix of streams.}

\item{Nglobal}{a (non-empty) integer vector specifying size of work items for use, with default value from global option 'Nglobal'}

\item{type}{a character string specifying "double" or "float" of random numbers, with default value from global option 'type'}

\item{verbose}{a logical value, if TRUE, print extra information, with default value from global option 'verbose'}
}
\value{
a vclVector or vclMatrix of uniform random numbers
}
\description{
Generate uniform random numbers parallely on a GPU
}
\details{
\code{type} specifies the precision type of random numbers. If GPU supports "double", default is "double", otherwise, default is "single"
}
\examples{
 
library('clrng')
currentDevice()
streams <- createStreamsGpu(8)
as.vector(runifGpu(5, streams))

# Change global options
as.matrix(runifGpu(c(2,2), streams))
}
