% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateStreamsGpu.R
\name{setBaseCreator}
\alias{setBaseCreator}
\title{setBaseCreator}
\usage{
setBaseCreator(initial = rep(12345, 6))
}
\arguments{
\item{initial}{Initial state of the first stream, length 6, recycled if shorter}
}
\value{
A stream object of class 'vclMatrix' on GPU
}
\description{
Set the initial seed of the package or stream
}
\examples{
setBaseCreator(c(111,222,333,444,555,666))
}
