% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{week}
\alias{week}
\title{Sample dataset: week}
\format{
A character vector with the names of the days of the week
}
\source{
Generated for demonstration purposes.
}
\usage{
data(week)
}
\description{
An example data set representing days of the week.
}
\keyword{datasets}
